# [SPARK-31426][SQL] Fix perf regressions of toJavaTimestamp/fromJavaTimestamp