@MaxGekk 2020-04-12
@cloud-fan @HyukjinKwon @dongjoon-hyun  Please, have a look at the PR.