@dongjoon-hyun 2020-04-14
Hi, @cloud-fan .
This seems to break `branch-3.0`. Could you take a look at that once more?
```
[ERROR] [Error] /home/runner/work/spark/spark/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/util/DateTimeUtils.scala:181: not found: value millisToMicros
700
[ERROR] one error found
```