@MaxGekk 2020-04-14
WDYT of merging https://github.com/apache/spark/pull/27618 to branch-3.0? Missing the changes in branch-3.0 causes problems already twice, see https://github.com/apache/spark/pull/27964#issuecomment-602153140